# pages/api-docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "4.0",
    "title": "18 Candleriggs",
    "description": "This is the Candleriggs Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "name": "Superminds",
      "email": "contact@superminds.dev"
    }
  },
  "paths": {
...
    "/api/banner/upload-desktop-image/{bannerId}": {
      "post": {
        "tags": [
          "Banners"
        ],
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Upload a banner desktop image",
        "description": "This endpoint is used to upload desktop banner image using bannerID.\n\n__How To__\n\n1. If you dont have a bannerID hit an endpoint **GET /api/banner** and get a valid bannerID\n2. Fill the bannerID.\n3. Select image to upload and click execute\n\n**Accepted image formats**\n1. png\n\n2. jpg\n\n3. jpeg\n\n **Accepted size and dimensions (width & height)** \n\nsize : under 1MB\n\nwidth: 1440px\n\nheight: 692px\n\n**The below table defines the HTTP Status codes that this API may return**\n\n  | Status Code      | Description | Reason                                   |\n  | ---------------- | ------------| -----------------------------------------|\n  | 201              | CREATED          | Banner desktop image created successfully.                       |\n  | 403              | FORBIDDEN | If user unauthenticated. |\n  | 404              | NOT FOUND   | If this bannerId not found.              |\n  | 422              | BAD REQUEST | If the bannerId is in an invalid format or body is not according to validations. |\n  | 500              | SERVER ERROR| If server error encountered.             |\n",
        "operationId": "bannerDesktopImageUpload",
        "parameters": [
          {
            "name": "bannerId",
            "in": "path",
            "required": true,
            "description": "The Id of the banner to upload desktop image",
            "schema": {
              "type": "string",
              "example": "63844c8bc1e969f2934f09f7"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Banner desktop image successfully uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 201
                    },
                    "msg": {
                      "type": "string",
                      "example": "Uploaded successfully"
                    },
                    "result": {
                      "allOf": [
                        {
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "63982444dd330c6edc3e3c9b"
                            }
                          }
                        },
                        {
                          "$ref": "#/components/schemas/Banner"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return400"
                }
              }
            },
            "description": "Data not received from user"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return403"
                }
              }
            },
            "description": "Forbidden: If user unauthenticated."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return404"
                }
              }
            },
            "description": "ObjectID not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return422"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return500"
                }
              }
            },
            "description": "Server side error."
          }
        }
      },
      "delete": {
        "tags": [
          "Banners"
        ],
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Delete the desktop image of a banner using bannerID",
        "description": "This endpoint is used to delete a desktop image of a banner using bannerID.\n\n__How To__\n\n1. If you dont have a bannerID hit an endpoint **GET /api/banner** and get a valid bannerID. \n2. Fill the bannerID and execute.\n\n**The below table defines the HTTP Status codes that this API may return**\n\n  | Status Code      | Description | Reason                                   |\n  | ---------------- | ------------| -----------------------------------------|\n  | 200              | OK          | Image deleted successfully.                       |\n  | 403              | FORBIDDEN | If user unauthenticated. |\n  | 404              | NOT FOUND   | If this bannerId not found.              |\n  | 422              | BAD REQUEST | If the bannerId is in an invalid format. |\n  | 500              | SERVER ERROR| If server error encountered.             |\n",
        "operationId": "bannerDesktopImageDELETE",
        "parameters": [
          {
            "name": "bannerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6399bf322f332f8babe66407"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "msg": {
                      "type": "string",
                      "example": "Deleted successfully"
                    }
                  }
                }
              }
            },
            "description": "success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return403"
                }
              }
            },
            "description": "Forbidden: If user unauthenticated."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return404"
                }
              }
            },
            "description": "BannerID not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return422"
                }
              }
            },
            "description": "Invalid Id provided"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return500"
                }
              }
            },
            "description": "Server side error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Return400": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 400
          },
          "msg": {
            "type": "string",
            "example": "Bad Request: {message}"
          }
        }
      },
      "Return403": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 403
          },
          "msg": {
            "type": "string",
            "example": "Forbidden: {message}"
          }
        }
      },
      "Return404": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 404
          },
          "msg": {
            "type": "string",
            "example": "Not Found: {message}"
          }
        }
      },
...
      "Return422": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 422
          },
          "msg": {
            "type": "string",
            "example": "Unprocessable Entity"
          },
          "errors": {
            "type": "array",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ValidationError"
                }
              ]
            }
          }
        }
      },
      "Return500": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 500
          },
          "msg": {
            "type": "string",
            "example": "Internal Server Error: {message}"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "example": "Provide Password"
          },
          "param": {
            "type": "string",
            "example": "password"
          },
          "location": {
            "type": "string",
            "example": "body"
          }
        }
      },
      "Banner": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BannerWithoutImages"
          },
          {
            "type": "object",
            "properties": {
              "desktopImage": {
                "type": "string",
                "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/image1"
              },
              "mobileImage": {
                "type": "string",
                "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/image2"
              }
            }
          }
        ]
      },
      "BannerWithoutImages": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "The Rocky Horror Tribute Show",
            "minLength": 1,
            "maxLength": 30
          },
          "link": {
            "type": "string",
            "example": "https://www.skiddle.com/whats-on-beta/Glasgow/18-Candleriggs-%28Formerly-Wild-Cabaret%29/The-Rocky-Horror-Tribute-Show-LIVE/35938181/"
          },
          "eventType": {
            "type": "string",
            "enum": [
              "Live",
              "Sold Out",
              "Cancelled",
              "Private Booking"
            ]
          },
          "startDate": {
            "type": "string",
            "example": "2002-07-15",
            "format": "full-date",
            "externalDocs": {
              "description": "SeeMore",
              "url": "https://www.rfc-editor.org/rfc/rfc3339#section-5.6"
            }
          },
          "endDate": {
            "type": "string",
            "example": "2002-07-15",
            "format": "full-date",
            "externalDocs": {
              "description": "SeeMore",
              "url": "https://www.rfc-editor.org/rfc/rfc3339#section-5.6"
            }
          }
        },
        "required": [
          "name",
          "link",
          "eventType",
          "startDate",
          "endDate"
        ]
      },
...
    }
  }
}